# ENH filter nan before constructing Python set