@glemaitre 2022-11-05
@thomasjpfan I don't know if you think that we would have a way to avoid the filtering if we can know in advance that we don't have missing values?