@adrinjalali 2022-11-07
Curious, how would your benchmark look like if you add something like a `if nan in list:` before the filter 🤔 it should avoid creation of a new list in case there is no nans.