@glemaitre 2022-11-07
Since `nan != nan` this would not work:

```python
In [1]: x = [float("nan"), 1, 2, float("nan")]

In [2]: float("nan") in x
Out[2]: False
```

You are too naive regarding `nan` values @adrinjalali ;)