@glemaitre 2022-11-07
somehow, the way I would like something that does this:

```python
In [4]: some_nan = False
   ...: for item in x:
   ...:     if math.isnan(item):
   ...:         some_nan = True
   ...:         break
   ...: 

In [5]: some_nan
Out[5]: True
```