@glemaitre 2022-11-07
Benchmark if the `if` statement will give the same timing in terms of execution time. It would avoid a copy of `(n_samples, 1)` however.