@adrinjalali 2022-11-07
So obviously it depends on where the `nan` happens to be in the list, but if there is no `nan`:

```py
In [18]: c = [x for x in range(100000)]

In [19]: %timeit any(map(math.isnan, c))
2.48 ms ± 34.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [20]: %timeit b = list(filterfalse(math.isnan, c))
3.19 ms ± 85.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

it seems we're gonna get some speedup, but not much?

I'd say since the speedup we get for datasets where we do have `nan`, this is worth it.