@betatim 2022-11-07
When I benchmark the following three cases I get these results:
```python
In [23]: c = [x for x in range(100000)]

In [24]: %timeit set(c)
1.35 ms ± 11.3 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [25]: %timeit _extract_set_with_possibly_nan(c)
33.2 ms ± 90.5 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [26]: %timeit Counter(c).keys()
3.05 ms ± 26.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
where `Counter` is from the collections module of the stdlib. Getting the keys is not equivalent to creating a set, but maybe it is close enough?

When I add one `np.nan` at the end `c = [x for x in range(100000)] + [np.nan]` then the time for `_extract_set_with_possible_nan` becomes something like 50 or 60ms.

Interestingly, the way that `Counter` is implemented is pretty simple:
```python
def _count_elements(mapping, iterable):
    'Tally elements from the iterable.'
    mapping_get = mapping.get
    for elem in iterable:
        mapping[elem] = mapping_get(elem, 0) + 1
```

Maybe that gives a pointer towards what to do?