@betatim 2022-11-07
What dataset/mix of NaN and normal values did you use @glemaitre ?