@glemaitre 2022-11-07
I use the example of the issue where you have 99% of missing values for a `10_000` vector.
I did not know about the feature of `Counter`. I am interesting to check how does it deal with `nan` then.