@glemaitre 2022-11-07
You will get the same issue:

```python
In [1]: x = [float("nan"), 1, 2, float("nan")]

In [2]: from collections import Counter

In [3]: Counter(x)
Out[3]: Counter({nan: 1, 1: 1, 2: 1, nan: 1})
```

So `nan != nan` and you get 2 entries. If you have a column with `9_000` NaNs, then the comparison will be expensive as with the set.