@betatim 2022-11-07
I didn't realise that there was more than one NaN :(

```python
In [25]: def count_elements(iterable):
    ...:     mapping = {}
    ...:     mapping_get = mapping.get
    ...:     np_nan = np.nan
    ...:     math_isnan = math.isnan
    ...:     numbers_Real = numbers.Real
    ...:     for elem in iterable:
    ...:         if isinstance(elem, numbers_Real) and math_isnan(elem):
    ...:             elem = np_nan
    ...:         mapping[elem] = mapping_get(elem, 0) + 1
    ...:     return set(mapping)
```

combines `is_scalar_nan` with the "one loop" approach from `Counter`. The ugly "inlining" of things like `numbers_Real` takes the speed from ~28ms to 21ms for me. So seems worth the hassle.